feat(schema) disallow empty aggregates on entities
Right know what the schema does with strings is something clever. it sets the
`len_min = 1` validator by default when `len_min` is not set. This does the
same for other aggregates.

This approach allows to set `len_min` to 0 in the metaschema for validating
schemas if needed, because don't go through the DB. This still forces
`len_min` to be `> 0` on DB entities via the `kong.db.schema.entity`
"superclass".

Also, this commits includes a nice error message when the user attempts
to set something with length = 0, `", use null to unset"`.